Добавляет сборку проекта. Разбивает код на модули
