+++ src/components/buttons.js
@@ -1,0 +1,5 @@
+const createShowMoreButton = () => {
+  return (`<button class="films-list__show-more">Show more</button>`);
+};
+
+export {createShowMoreButton};
