+++ src/components/constants.js
@@ -1,0 +1,4 @@
+const MAIN_FILMS_COUNT = 5;
+const OTHER_FILMS_COUNT = 2;
+
+export {MAIN_FILMS_COUNT, OTHER_FILMS_COUNT};
